+++ chaospy/distributions/baseclass/copula.py
@@ -49,6 +49,7 @@ together in `chaospy`.
 import numpy
 import chaospy
 
+from .index import Index
 from .distribution import Distribution
 
 
@@ -75,10 +76,16 @@ class Copula(Distribution):
             dependencies=dependencies,
             rotation=rotation,
             repr_args=repr_args,
+            index_cls=CopulaIndex,
         )
 
-    def _cdf(self, x, dist, trans, cache):
-        output = dist._get_fwd(x, cache=cache)
+    def _cdf(self, xloc, dist, trans, cache):
+        out = numpy.zeros(xloc.shape)
+        for idx in self._rotation:
+            xloc_ = xloc[idx].reshape(1, -1)
+            out[idx] = self[int(idx)]._get_fwd(xloc_, cache)
+        return out
+        output = dist._get_fwd(xloc, cache=cache)
         output = trans._get_fwd(output, cache=cache)
         return output
 
@@ -89,6 +96,11 @@ class Copula(Distribution):
         return dist._get_upper(cache=cache)
 
     def _ppf(self, qloc, dist, trans, cache):
+        out = numpy.zeros(qloc.shape)
+        for idx in self._rotation:
+            qloc_ = qloc[idx].reshape(1, -1)
+            out[idx] = self[int(idx)]._get_inv(qloc_, cache)
+        return out
         qloc = trans._get_inv(qloc, cache=cache)
         xloc = dist._get_inv(qloc, cache=cache)
         return xloc
@@ -98,8 +110,46 @@ class Copula(Distribution):
         return trans._get_pdf(dist._get_fwd(x, cache=cache), cache=cache)*density
 
     def _mom(self, x, dist, trans, cache):
-        raise chaospy.UnsupportedFeature(
-            "Joint distribution with dependencies not supported.")
+        raise chaospy.UnsupportedFeature("Copula not supported.")
+
+    def _ttr(self, x, dist, trans, cache):
+        raise chaospy.UnsupportedFeature("Copula not supported.")
 
     def _cache(self, dist, trans, cache):
+        return self
+
+
+class CopulaIndex(Index):
+
+    def __init__(self, parent, conditions=()):
+        assert isinstance(parent, Copula)
+        super(CopulaIndex, self).__init__(
+            parent=parent, conditions=conditions)
+
+    def _ppf(self, qloc, idx, parent, conditions, cache):
+        parameters = parent.get_parameters(cache, assert_numerical=False)
+        qloc = parameters["trans"][int(idx)]._get_inv(qloc, cache=cache)
+        xloc = parameters["dist"][int(idx)]._get_inv(qloc, cache=cache)
+        return xloc
+
+    def _cdf(self, xloc, idx, parent, conditions, cache):
+        parameters = parent.get_parameters(cache, assert_numerical=False)
+        output = parameters["dist"][int(idx)]._get_fwd(xloc, cache=cache)
+        output = parameters["trans"][int(idx)]._get_fwd(output, cache=cache)
+        return output
+
+    def _pdf(self, xloc, idx, parent, conditions, cache):
+        parameters = parent.get_parameters(cache, assert_numerical=False)
+        density = parameters["dist"][int(idx)]._get_pdf(xloc, cache=cache.copy())
+        zloc = parameters["dist"][int(idx)]._get_fwd(xloc, cache=cache)
+        output = parameters["trans"][int(idx)]._get_pdf(zloc, cache=cache)*density
+        return output
+
+    def _mom(self, kloc, idx, parent, conditions, cache):
+        raise chaospy.UnsupportedFeature("Copula not supported.")
+
+    def _ttr(self, kloc, idx, parent, conditions, cache):
+        raise chaospy.UnsupportedFeature("Copula not supported.")
+
+    def _cache(self, idx, parent, conditions, cache):
         return self